requirements.txt
# dbms products
# env products_operation  

psycopg2==2.8.6
Django==3.1.2
Pillow==8.0.1
djangorestframework==3.12.2
django-rest-swagger==2.2.0
drf-yasg==1.20.0
gunicorn==20.0.4
django-heroku==0.3.1